backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv



pymysql

pandas
librosa
scikit-learn